refactor(persistence): decompose alerts.rs into focused modules [OPE-387]
Split monolithic alerts.rs (719 LOC) into alerts/ directory with:
- mod.rs: module declarations and public re-exports
- types.rs: AlertAction, AlertMetric, AlertCondition, AlertRule, and related types
- storage.rs: AlertStore CRUD, history, and metrics queries
- tests.rs: all 19 existing tests (unchanged)

All public API exports preserved. cargo check --workspace and cargo test pass.
